% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTGM.R
\name{htgmM}
\alias{htgmM}
\title{htgmM}
\usage{
htgmM(l, fdrThresh)
}
\arguments{
\item{l}{list of return values of GoMiner()}

\item{fdrThresh}{numerical acceptance threshold for fdr}
}
\value{
returns numeric matrix m containing FDR values
}
\description{
generate FDR matrix of id versus cat
}
\examples{
# load("data/x_l.RData")
m<-htgmM(x_l,.1)

}
