% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTGM.R
\name{hyperGenes}
\alias{hyperGenes}
\title{hyperGenes}
\usage{
hyperGenes(l, dir)
}
\arguments{
\item{l}{return value of GoMiner()}

\item{dir}{character string containing path name of results directory}
}
\value{
returns no value but has side effect of populating subdirectory of
hyperlinked gene lists
}
\description{
populate subdirectory of hyperlinked gene lists
}
\examples{
# x_l<-load("data/x_l.RData")
dir<-tempdir()
print(dir)
hyperGenes(x_l,dir)
 
}
