% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTGM.R
\name{hyperlinks}
\alias{hyperlinks}
\title{hyperlinks}
\usage{
hyperlinks(s, rownames, colnames)
}
\arguments{
\item{s}{character path name of the file containing the HTGM svg}

\item{rownames}{character vector of row names}

\item{colnames}{character vector of column names}
}
\value{
returns the path name of the file containing the hyperlinked HTGM svg
}
\description{
driver to add gene list hyperlinks to the HTGM heatmap
}
\examples{
#load("data/x_rn.RData")
#load("data/x_cn.RData")
#load("data/x_svg.RData")
s<-system.file("extdata","x_htgm.svg",package="HTGM")
hyperlinkedFileName<-hyperlinks(s,x_rn,x_cn)
print("hyperlinkedFileName")
print(hyperlinkedFileName)

}
