% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTGM.R
\name{pasteHyperlinks}
\alias{pasteHyperlinks}
\title{pasteHyperlinks}
\usage{
pasteHyperlinks(str, c1, c2)
}
\arguments{
\item{str}{character a line from the svg that is to have a hyperlink inserted}

\item{c1}{character list of row names}

\item{c2}{character list of column names}
}
\value{
returns a line of code to insert into svg
}
\description{
add gene list hyperlinks to the HTGM heatmap
}
\examples{
#load("data/x_svgr.RData")
#load("data/x_rnr.RData")
#load("data/x_cnc.RData")
hl<-pasteHyperlinks(x_svgr,x_rnr,x_cnc)

}
